# app/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Church Management API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/register": {
      "post": {
        "tags": ["user"],
        "summary": "Registra un nuevo usuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado"
          },
          "400": {
            "description": "Solicitud incorrecta"
          },
          "409": {
            "description": "El usuario ya existe"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Inicia sesión con un usuario existente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    },
    "/api/whoami": {
      "get": {
        "tags": ["auth"],
        "summary": "Obtiene detalles del usuario autenticado",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user_details": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string"
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/api/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Refresca el token de acceso",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token de acceso refrescado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/api/logout": {
      "get": {
        "tags": ["auth"],

        "summary": "Cierra la sesión del usuario actual",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token removido exitosamente"
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["user"],

        "summary": "Obtiene una lista de todos los usuarios (solo admin)",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/user": {
      "get": {
        "tags": ["user"],

        "summary": "Obtiene el usuario autenticado",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/scanner": {
      "post": {
        "tags": ["event_detail"],
        "summary": "Crea un nuevo detalle de evento desde el escáner",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "guest_id": {
                    "type": "integer"
                  },
                  "observaciones": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "integer"
                  }
                },
                "required": ["guest_id"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Detalle de evento creado"
          },
          "400": {
            "description": "Solicitud incorrecta"
          },
          "404": {
            "description": "No se encontró un evento con estado 0"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/event_details": {
      "get": {
        "tags": ["event_detail"],
        "summary": "Obtiene una lista de todos los detalles de eventos",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de detalles de eventos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDetail"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "post": {
        "tags": ["event_detail"],
        "summary": "Crea un nuevo detalle de evento",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDetail"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Detalle de evento creado"
          },
          "400": {
            "description": "Solicitud incorrecta"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/event_details/{id}": {
      "get": {
        "tags": ["event_detail"],
        "summary": "Obtiene un detalle de evento por ID",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "404": {
            "description": "Detalle de evento no encontrado"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "put": {
        "tags": ["event_detail"],
        "summary": "Actualiza un detalle de evento por ID",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDetail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detalle de evento actualizado"
          },
          "404": {
            "description": "Detalle de evento no encontrado"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "delete": {
        "tags": ["event_detail"],
        "summary": "Elimina un detalle de evento por ID",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Detalle de evento eliminado"
          },
          "404": {
            "description": "Detalle de evento no encontrado"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "JWTAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "EventDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "hora": {
            "type": "string",
            "format": "date-time"
          },
          "event_id": {
            "type": "integer"
          },
          "guest_id": {
            "type": "integer"
          },
          "observaciones": {
            "type": "string"
          },
          "user_id": {
            "type": "integer"
          }
        },
        "required": ["hora", "event_id", "guest_id", "user_id"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["admin", "user"]
            }
          }
        },
        "required": ["email", "password", "roles"]
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"]
      },
      "Token": {
        "type": "object",
        "properties": {
          "access": {
            "type": "string"
          },
          "refresh": {
            "type": "string"
          }
        }
      }
    }
  }
}